abstract out createNotebook
